fixing redirection
